+++ src/test/java/com/github/edipermadi/security/blobfish/pool/BlobPoolBuilderTest.java
@@ -119,6 +119,25 @@ public final class BlobPoolBuilderTest extends AbstractTest {
     }
 
     @Test(dependsOnMethods = {"testImportPayloadByPrivateKey"})
+    public void testFindTags() throws SQLException {
+        final String[] keywords = {"fish", "deep", "sea"};
+
+        for (final String keyword : keywords) {
+            log("looking for keyword '%s'", keyword);
+            boolean empty = false;
+            for (int page = 1; !empty; page++) {
+                final Map<UUID, String> tags = blobPool.findTags(keyword, page, 10);
+                for (final Map.Entry<UUID, String> entry : tags.entrySet()) {
+                    log("  found entry");
+                    log("    uuid : %s", entry.getKey());
+                    log("    tag  : %s", entry.getValue());
+                }
+                empty = tags.isEmpty();
+            }
+        }
+    }
+
+    @Test(dependsOnMethods = {"testImportPayloadByPrivateKey"})
     public void testCreateBlob() throws SQLException, IOException {
         final String content = RandomStringUtils.randomAlphanumeric(256);
 
@@ -221,7 +240,8 @@ public final class BlobPoolBuilderTest extends AbstractTest {
 
         /* update payload */
         try (final ByteArrayInputStream bais = new ByteArrayInputStream(contentNew.getBytes(StandardCharsets.US_ASCII))) {
-            blobPool.updateBlobPayload(blobId, bais);
+            final boolean updated = blobPool.updateBlobPayload(blobId, bais);
+            Assert.assertTrue(updated);
         }
 
         /* check content */
@@ -365,6 +385,26 @@ public final class BlobPoolBuilderTest extends AbstractTest {
                 log("    name : %s", entry.getValue());
             }
             empty = recipients.isEmpty();
+        }
+    }
+
+    @Test(dependsOnMethods = {"testCreateRecipient"})
+    public void testFindRecipient() throws SQLException {
+        Assert.assertNotNull(blobPool);
+
+        final String[] keywords = {"edipermadi", "github", "blobfish"};
+        for(final String keyword:keywords){
+            log("finding recipient with keyword '%s'", keyword);
+            boolean empty = false;
+            for (int page = 1; !empty; page++) {
+                final Map<UUID, String> recipients = blobPool.findRecipient(keyword, page, 10);
+                for (Map.Entry<UUID, String> entry : recipients.entrySet()) {
+                    log("  found recipient");
+                    log("    uuid : %s", entry.getKey());
+                    log("    name : %s", entry.getValue());
+                }
+                empty = recipients.isEmpty();
+            }
         }
     }
 
@@ -579,6 +619,43 @@ public final class BlobPoolBuilderTest extends AbstractTest {
                 }
             }
             empty = blobs.isEmpty();
+        }
+    }
+
+    @Test(dependsOnMethods = {"testImportPayloadByPrivateKey"})
+    public void testFindBlobs() throws SQLException, IOException {
+        final String[] keywords = {"amused", "annoyed", "free", "glasses", "happy", "sad", "surprised"};
+
+        for(final String keyword : keywords) {
+            boolean empty = false;
+            log("listing blobs with keyword '%s'", keyword);
+            for (int page = 1; !empty; page++) {
+                final Map<UUID, Blob.SimplifiedMetadata> blobs = blobPool.findBlobs(keyword, page, 10);
+                for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : blobs.entrySet()) {
+                    final Blob.SimplifiedMetadata metadata = entry.getValue();
+                    final Map<UUID, String> tags = blobPool.getBlobTags(entry.getKey());
+                    final byte[] payload = blobPool.getBlobPayload(entry.getKey());
+
+                    final Set<String> tagValues = new HashSet<>();
+                    for (final Map.Entry<UUID, String> tag : tags.entrySet()) {
+                        tagValues.add(tag.getValue());
+                    }
+
+                    log("  found entry");
+                    log("    uuid      : %s", entry.getKey());
+                    log("    mime-type : %s", metadata.getMimeType());
+                    log("    path      : %s", metadata.getPath());
+                    log("    tags      : %s", Joiner.on(", ").join(tagValues));
+
+                    final File file = new File(metadata.getPath());
+                    final File dir = new File("target");
+                    final File outFile = new File(dir, "pool-" + file.getName());
+                    try (final FileOutputStream fos = new FileOutputStream(outFile)) {
+                        fos.write(payload);
+                    }
+                }
+                empty = blobs.isEmpty();
+            }
         }
     }